Implement keyboard handling for popups on iOS and Android
